perf(checker): delegation-tree memo for completed cross-arena sentinel results fixes drizzle-orm livelock
The drizzle-orm benchmark canary pegged one core for >900s with zero
diagnostics (issue #13041): every type-reference occurrence of a
cross-file symbol whose delegation completes with a sentinel
(ERROR/UNKNOWN) re-ran the full child-checker pipeline, because the
shared DefinitionStore cross-file buckets refuse sentinel writes (an
in-progress sentinel must not poison other checkers). One alias such as
SelectedFields re-ran 2,693 identical full delegations in a 73-file
subset, all completing with the same ERROR.

Add CrossArenaSessionMemo (embedded in CrossFileDelegationCache):
completed sentinel outcomes keyed by (owner_file_idx, raw SymbolId,
in-progress-context fingerprint), scoped to one outermost delegation
tree (cleared when a delegation entry point roots a new tree at
cross-arena depth 0), shared via Arc with transient child checkers.
Replay happens only for repeats under the identical in-progress
resolution context within the producing tree, where recomputation is
the same deterministic function. Successful results stay on the gated
shared-store buckets.

drizzle-orm: timeout >900s -> ~390s (completes). 73-file subset: 38.1s
-> 19.0s, full-work delegation misses 6,715 -> 798, diagnostics
byte-identical. Canary error counts (zod/comlink/msw/ofetch/ts-rest/
kysely/valibot) all match own-baseline within the canaries' own
run-to-run determinism envelope. Filtered tsz-checker nextest failure
set byte-identical to origin/main.

Also splits the alias-shortcut helpers out of cross_file.rs to respect
the 2000-line shard cap.

Fixes #13041
